+++ R/cor_test.R
@@ -8,7 +8,7 @@
 #'   set to 0.95 (95\% CI).
 #' @param method A character string indicating which correlation coefficient is
 #'   to be used for the test. One of \code{"pearson"} (default),
-#'   \code{"kendall"}, \code{"spearman"}, \code{"biserial"},
+#'   \code{"kendall"}, \code{"spearman"} (but see also the \code{robust} argument), \code{"biserial"},
 #'   \code{"polychoric"}, \code{"tetrachoric"}, \code{"biweight"},
 #'   \code{"distance"}, \code{"percentage"} (for percentage bend correlation),
 #'   \code{"blomqvist"} (for Blomqvist's coefficient), \code{"hoeffding"} (for
@@ -37,15 +37,21 @@
 #'   \code{BayesFactor::correlationBF} function.
 #' @param bayesian_ci_method,bayesian_test See arguments in
 #'   \code{\link[=parameters]{model_parameters}} for \code{BayesFactor} tests.
-#' @param robust If \code{TRUE}, will rank-transform the variables prior to
-#'   estimating the correlation. Note that, for instance, a Pearson's
+#' @param ranktransform If \code{TRUE}, will rank-transform the variables prior to
+#'   estimating the correlation, which is one way of making the analysis more
+#'   resistant to extreme values (outliers). Note that, for instance, a Pearson's
 #'   correlation on rank-transformed data is equivalent to a Spearman's rank
 #'   correlation. Thus, using \code{robust=TRUE} and \code{method="spearman"} is
-#'   redundant. Nonetheless, it is an easy way to increase the robustness of the
-#'   correlation (as well as obtaining Bayesian Spearman rank Correlations).
-#' @param winsorize Either \code{FALSE} or a number between 0 and 1 (e.g.,
-#'   \code{0.2}) that corresponds to the threshold of desired
-#'   \code{\link[=winsorize]{winsorization}}.
+#'   redundant. Nonetheless, it is an easy option to increase the robustness of the
+#'   correlation as well as flexible way to obtain Bayesian or multilevel
+#'   Spearman-like rank correlations.
+#' @param robust Old name for \code{ranktransform}. Will be removed in subsequent
+#'   versions, so better to use \code{ranktransform} which is more explicit about
+#'   what it does.
+#' @param winsorize Another way of making the correlation more "robust" (i.e.,
+#'   limiting the impact of extreme values). Can be either \code{FALSE} or a
+#'   number between 0 and 1 (e.g., \code{0.2}) that corresponds to the desired
+#'   threshold. See the \code{\link[=winsorize]{winsorize()}} function for more details.
 #' @param verbose Toggle warnings.
 #' @param ... Additional arguments (e.g., \code{alternative}) to be passed to
 #'   other methods. See \code{stats::cor.test} for further details.
@@ -97,8 +103,8 @@
 #' }
 #'
 #' # Robust (these two are equivalent)
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "pearson", robust = TRUE)
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "spearman", robust = FALSE)
+#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "spearman")
+#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "pearson", ranktransform = TRUE)
 #'
 #' # Winsorized
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", winsorize = 0.2)
@@ -125,10 +131,17 @@ cor_test <- function(data,
                      partial = FALSE,
                      partial_bayesian = FALSE,
                      multilevel = FALSE,
-                     robust = FALSE,
+                     ranktransform = FALSE,
+                     robust = NULL,
                      winsorize = FALSE,
                      verbose = TRUE,
                      ...) {
+
+  # Deprecation warnings
+  if (!is.null(robust)) {
+    warning("The 'robust' argument is deprecated in favour of 'ranktransform' (more explicit). Please use the latter instead to remove this warning.")
+    ranktransform <- robust
+  }
 
   # Sanity checks
   if (!x %in% names(data) | !y %in% names(data)) {
@@ -151,19 +164,19 @@ cor_test <- function(data,
 
   # Winsorize
   if (!isFALSE(winsorize) && !is.null(winsorize)) {
-    # set default
+    # set default (if not specified)
     if (isTRUE(winsorize)) {
-      winsorize <- .1
+      winsorize <- .2
     }
 
     # winsorization would otherwise fail in case of NAs present
-    data <- stats::na.omit(data)
+    # data <- stats::na.omit(data)
 
-    data[c(x, y)] <- winsorize(data[c(x, y)], threshold = winsorize, verbose = verbose)
+    data <- as.data.frame(winsorize(stats::na.omit(data[c(x, y)]), threshold = winsorize, verbose = verbose))
   }
 
-  # Robust
-  if (robust) {
+  # Rank transform (i.e., "robust")
+  if (ranktransform) {
     data[c(x, y)] <- effectsize::ranktransform(data[c(x, y)], sign = FALSE, method = "average")
   }