+++ R/winsorize.R
@@ -1,6 +1,14 @@
 #' Winsorize data
 #'
-#' Winsorization of data.
+#' Winsorizing or winsorization is the transformation of statistics by limiting
+#' extreme values in the statistical data to reduce the effect of possibly
+#' spurious outliers. The distribution of many statistics can be heavily
+#' influenced by outliers. A typical strategy is to set all outliers (values
+#' beyond a certain threshold) to a specified percentile of the data; for
+#' example, a 90\% winsorization would see all data below the 5th percentile set
+#' to the 5th percentile, and data above the 95th percentile set to the 95th
+#' percentile. Winsorized estimators are usually more robust to outliers than
+#' their more standard forms.
 #'
 #' @param data Dataframe or vector.
 #' @param threshold The amount of winsorization.
@@ -19,8 +27,8 @@ winsorize <- function(data, ...) {
 
 
 #' @export
-winsorize.factor <- function(data, threshold = 0.2, ...) {
-  stats::na.omit(data)
+winsorize.factor <- function(data, ...) {
+  data
 }
 
 #' @export
@@ -31,7 +39,7 @@ winsorize.logical <- winsorize.factor
 
 #' @export
 winsorize.data.frame <- function(data, threshold = 0.2, verbose = TRUE, ...) {
-  sapply(stats::na.omit(data), winsorize, threshold = threshold, verbose = verbose)
+  sapply(data, winsorize, threshold = threshold, verbose = verbose)
 }
 
 #' @rdname winsorize
@@ -44,7 +52,6 @@ winsorize.numeric <- function(data, threshold = 0.2, verbose = TRUE, ...) {
     return(data)
   }
 
-  data <- stats::na.omit(data)
   y <- sort(data)
   n <- length(data)
   ibot <- floor(threshold * n) + 1